requirements.txt
atlassian-python-api==3.41.16
beautifulsoup4==4.12.3
datasets
fastapi==0.115.5
fastembed
html2text==2024.2.26
llama-index==0.11.23
llama-index-embeddings-fastembed==0.2.0
llama-index-embeddings-nvidia
llama-index-llms-gemini
llama-index-llms-openai==0.2.16
llama-index-readers-confluence==0.2.1
llama-index-readers-file==0.3.0
llama-index-readers-github
llama-index-vector-stores-qdrant==0.3.3
pdf2image==1.17.0
Pillow==10.4.0
pydantic==2.9.2
pytest
pytesseract==0.3.13
qdrant-client==1.12.1
requests==2.32.3
sentence-transformers==2.7.0
torch
transformers==4.42.4
uvicorn==0.32.0
